docs: Add contributing/architecture docs
