+++ internal/rules/policies/cloud/policies/aws/rds/enable_iam_auth.rego
@@ -1,6 +1,6 @@
 # METADATA
 # title: "RDS IAM Database Authentication Disabled"
-# description: "Ensure IAM Database Authentication is enabled for RDS database instances to manage database acces"
+# description: "Ensure IAM Database Authentication is enabled for RDS database instances to manage database access"
 # scope: package
 # schemas:
 # - input: schema.input
